Add getFrameCount and name of Simulation.
Summary:
Simulation now provides getFrameCount method, which returns number
of all frames of the simulation counted until the moment of that
method call (it doesnt need to be the actual number of all frames
of the simulation).
Also, simulation now has a name, which should be displayed and used
in serialization.

Test Plan: Darth Vader

Reviewers: PhazzaT

Subscribers: bartekz450

Differential Revision: http://zpp.cloudapp.net/D3