Add proper colouring of atoms.
From now on, while loading a simulation from binary file,
a colors file is required to be in the same directory.
The colors file must be named same as the simulation file,
except of the extension, which should be pdb.meta.
The mentioned file is generated by motions, while creating
PDB file.